I dropped a class!
This was actually kind of hard for me to do as I kind of wanted to get
all my courses out of the way early. That said I think it's best to take
fewer classes and put more focus into them, and my research. Giving
myself a pat on the back for not overworking myself this semester lol